Moved heartbeat package out to /pkg directory as discussed via mailing-list. Removed creation of dual go routine to manage the heartbeat as unneccesary.
