+++ cmd/core-data/main.go
@@ -28,7 +28,7 @@ import (
 
 	"github.com/tsconn23/edgex-go"
 	"github.com/tsconn23/edgex-go/core/data"
-	"github.com/tsconn23/edgex-go/support/heartbeat"
+	"github.com/tsconn23/edgex-go/pkg/heartbeat"
 	"github.com/tsconn23/edgex-go/support/logging-client"
 )
